Make the usage of delete_all_objects clearer
Using `delete_all_objects` to delete a large number of objects (like a folder with millions objects) is an important use case and probably the reason this function exists, though it's not clear how it can be used this way - I had to try it just cause it would make sense (and it worked, which is awesome :)). 

So here I want to make it clearer that delete_all_objects not only accepts the list of binary object IDs but can also work with a stream. 

As for spec adjustment, I can also imagine creating a custom type and using it here, something like `@type objects :: [binary | {binary, binary}, ...] | Enum.t`, important part is: this function does not strictly require a list of binaries.